% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.R
\name{summarise_results}
\alias{summarise_results}
\title{Summarise bias and dispersion (MADm) across simulated fits}
\usage{
summarise_results(results)
}
\arguments{
\item{results}{Output from \code{run_model()}.}
}
\value{
Data frame with only parameters that were estimated in the Stan model.
}
\description{
Summarise bias and dispersion (MADm) across simulated fits
}
\examples{
\dontrun{
sim <- simulate_data(ind = 10, partners = 2, repeats = 1, iterations = 1)
res <- run_model(sim, model = "Trait.stan", iter = 100, cores = 2)
summary <- summarise_results(res)
}
}
