% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Fit one of the available Stan models to simulated datasets}
\usage{
run_model(sim, model = NULL, iter = 2000, seed = 1234, cores = 1)
}
\arguments{
\item{sim}{Output from \code{simulate_data()}.}

\item{model}{Name of the Stan model to use (choose from available options).}

\item{iter}{Number of iterations per chain (default = 1000).}

\item{seed}{Random seed for reproducibility.}

\item{cores}{Number of CPU cores (used if cmdstanr or rstan is available).}
}
\value{
A list of fitted model summaries, one per dataset.
}
\description{
Fit one of the available Stan models to simulated datasets
}
\examples{
\donttest{
sim <- simulate_data(B_0 = 1, Valpha=0.2, Vepsilon = 0.1)
res <- run_model(sim, model = "VP.stan", iter=2000, cores = 2)
}
}
